+++ woodwork/logical_types.py
@@ -2,7 +2,7 @@ import re
 import warnings
 
 import pandas as pd
-import pandas.api.types as pdtypes
+from pandas.api import types as pdtypes
 
 from woodwork.accessor_utils import _is_dask_series, _is_spark_series
 from woodwork.config import config
@@ -13,6 +13,7 @@ from woodwork.exceptions import (
 )
 from woodwork.type_sys.utils import _get_specified_ltype_params
 from woodwork.utils import (
+    _check_data_type_equality,
     _infer_datetime_format,
     _is_valid_latlong_series,
     _reformat_to_latlong,
@@ -40,7 +41,7 @@ class LogicalType(object, metaclass=LogicalTypeMetaClass):
     """Base class for all other Logical Types"""
 
     type_string = ClassNameDescriptor()
-    primary_dtype = "string"
+    primary_dtype = "string[pyarrow]"
     backup_dtype = None
     standard_tags = set()
 
@@ -63,7 +64,7 @@ class LogicalType(object, metaclass=LogicalTypeMetaClass):
     def transform(self, series):
         """Converts the series dtype to match the logical type's if it is different."""
         new_dtype = self._get_valid_dtype(type(series))
-        if new_dtype != str(series.dtype):
+        if new_dtype != str(series.dtype) and new_dtype != series.dtype:
             # Update the underlying series
             try:
                 series = series.astype(new_dtype)
@@ -76,7 +77,7 @@ class LogicalType(object, metaclass=LogicalTypeMetaClass):
         specific validation, as required. When the series' dtype does not match the logical types' required dtype,
         raises a TypeValidationError."""
         valid_dtype = self._get_valid_dtype(type(series))
-        if valid_dtype != str(series.dtype):
+        if not _check_data_type_equality(valid_dtype, str(series.dtype)):
             raise TypeValidationError(
                 f"Series dtype '{series.dtype}' is incompatible with {self.type_string} dtype."
             )
@@ -92,7 +93,7 @@ class Address(LogicalType):
             ['26387 Russell Hill, Dallas, TX 34521', '54305 Oxford Street, Seattle, WA 95132']
     """
 
-    primary_dtype = "string"
+    primary_dtype = "string[pyarrow]"
 
 
 class Age(LogicalType):
@@ -216,7 +217,7 @@ class Categorical(LogicalType):
     """
 
     primary_dtype = "category"
-    backup_dtype = "string"
+    backup_dtype = "string[pyarrow]"
     standard_tags = {"category"}
 
     def __init__(self, encoding=None):
@@ -237,7 +238,7 @@ class CountryCode(LogicalType):
     """
 
     primary_dtype = "category"
-    backup_dtype = "string"
+    backup_dtype = "string[pyarrow]"
     standard_tags = {"category"}
 
 
@@ -252,7 +253,7 @@ class CurrencyCode(LogicalType):
     """
 
     primary_dtype = "category"
-    backup_dtype = "string"
+    backup_dtype = "string[pyarrow]"
     standard_tags = {"category"}
 
 
@@ -374,7 +375,7 @@ class EmailAddress(LogicalType):
              "team@example.com"]
     """
 
-    primary_dtype = "string"
+    primary_dtype = "string[pyarrow]"
 
     def validate(self, series, return_invalid_values=False):
         """Validates email address values based on the regex in the config.
@@ -401,7 +402,7 @@ class Filepath(LogicalType):
              "/tmp"]
     """
 
-    primary_dtype = "string"
+    primary_dtype = "string[pyarrow]"
 
 
 class PersonFullName(LogicalType):
@@ -416,7 +417,7 @@ class PersonFullName(LogicalType):
              "James Brown"]
     """
 
-    primary_dtype = "string"
+    primary_dtype = "string[pyarrow]"
 
 
 class IPAddress(LogicalType):
@@ -431,7 +432,7 @@ class IPAddress(LogicalType):
              "2001:0db8:0000:0000:0000:ff00:0042:8329"]
     """
 
-    primary_dtype = "string"
+    primary_dtype = "string[pyarrow]"
 
 
 class LatLong(LogicalType):
@@ -496,7 +497,7 @@ class NaturalLanguage(LogicalType):
              "When will humans go to mars?"]
     """
 
-    primary_dtype = "string"
+    primary_dtype = "string[pyarrow]"
 
 
 class Unknown(LogicalType):
@@ -511,7 +512,7 @@ class Unknown(LogicalType):
 
     """
 
-    primary_dtype = "string"
+    primary_dtype = "string[pyarrow]"
 
 
 class Ordinal(LogicalType):
@@ -531,7 +532,7 @@ class Ordinal(LogicalType):
     """
 
     primary_dtype = "category"
-    backup_dtype = "string"
+    backup_dtype = "string[pyarrow]"
     standard_tags = {"category"}
 
     def __init__(self, order=None):
@@ -577,7 +578,7 @@ class Ordinal(LogicalType):
 
 class PhoneNumber(LogicalType):
     """Represents Logical Types that contain numeric digits and characters
-    representing a phone number
+    representing a phone number.
 
     Examples:
         .. code-block:: python
@@ -587,7 +588,20 @@ class PhoneNumber(LogicalType):
              "5551235495"]
     """
 
-    primary_dtype = "string"
+    primary_dtype = "string[pyarrow]"
+
+    def validate(self, series, return_invalid_values=False):
+        """Validates PhoneNumber values based on the regex in the config.
+        By default, this validates US/Canada-based phone numbers.
+
+        Args:
+            series (Series): Series of phone number values.
+            return_invalid_values (bool): Whether or not to return invalid phone numbers.
+
+        Returns:
+            Series: If return_invalid_values is True, returns invalid phone numbers.
+        """
+        return _regex_validate("phone_inference_regex", series, return_invalid_values)
 
 
 class SubRegionCode(LogicalType):
@@ -603,7 +617,7 @@ class SubRegionCode(LogicalType):
     """
 
     primary_dtype = "category"
-    backup_dtype = "string"
+    backup_dtype = "string[pyarrow]"
     standard_tags = {"category"}
 
 
@@ -633,7 +647,7 @@ class URL(LogicalType):
              "example.com"]
     """
 
-    primary_dtype = "string"
+    primary_dtype = "string[pyarrow]"
 
     def validate(self, series, return_invalid_values=False):
         """Validates URL values based on the regex in the config.
@@ -661,7 +675,7 @@ class PostalCode(LogicalType):
     """
 
     primary_dtype = "category"
-    backup_dtype = "string"
+    backup_dtype = "string[pyarrow]"
     standard_tags = {"category"}
 
 
@@ -680,7 +694,7 @@ _NULLABLE_PHYSICAL_TYPES = {
     "float64",
     "float128",
     "object",
-    "string",
+    "string[pyarrow]",
     "timedelta64[ns]",
 }
 
@@ -712,6 +726,7 @@ def _regex_validate(regex_key, series, return_invalid_values):
             type_string = {
                 "url_inference_regex": "url",
                 "email_inference_regex": "email address",
+                "phone_inference_regex": "phone number",
             }[regex_key]
 
             info = f"Series {series.name} contains invalid {type_string} values. "